add repositories and implement with default implementation

+++ src/domain/Entitys/Shape_Property.cs
@@ -5,9 +5,9 @@ namespace domain.Entitys
 {
     public class Shape_Property
     {
-        public int ShapeId { get; set; }
+        public long ShapeId { get; set; }
         public Shape Shape { get; set; }
-        public int PropertyId { get; set; }
+        public long PropertyId { get; set; }
         public Property Property { get; set; }
         public float CompatibilityRatio { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
